+++ src/types/api.type.ts
@@ -29,3 +29,13 @@ export interface ReviewWriteDTO {
   badFeedback: string[];
   ban: boolean;
 }
+
+export interface LoginDTO {
+  success: boolean;
+  token: string;
+  refreshToken: string;
+  nickname: string;
+  userId: string;
+  rtokenExpireTime: string;
+  tokenExpireTime: string;
+}
