--- src/types/dto/login.type.ts
@@ -1,9 +1,0 @@
-export interface LoginDTO {
-  success: boolean;
-  token: string;
-  refreshToken: string;
-  nickname: string;
-  userId: string;
-  rtokenExpireTime: string;
-  tokenExpireTime: string;
-}
